requirements.txt
colorama==0.4.6
h5py==3.8.0
Logbook==1.6.0
matplotlib==3.5.3
munch==4.0.0
nltk==3.8.1
numba==0.56.4
numpy==1.21.6
pandas==1.3.5
Pillow==10.3.0
pycocotools==2.0.7
PyYAML==6.0.1
scikit_learn==1.0.2
torch==1.13.1
torchtext==0.14.1
torchvision==0.14.1
tqdm==4.66.1
transformers==4.30.2
ujson==5.7.0
geomloss==0.2.6
tensorboard==2.20.0